+++ controllers/contacts/getListContacts.js
@@ -1,0 +1,16 @@
+const contacts = require('../../models/contacts');
+const { RequestError } = require('../../helpers');
+
+const getListContacts = async (req, res, next) => {
+  try {
+      const result = await contacts.listContacts()
+      if (!result) {
+          throw RequestError(404, error.message)
+      }
+    res.json(result)
+  } catch (error) {
+    next(error)
+  }
+}
+
+module.exports = getListContacts;
