+++ controllers/contacts/index.js
@@ -1,0 +1,13 @@
+const getListContacts = require('./getListContacts');
+const getById = require('./getById');
+const add = require('./add');
+const remove = require('./remove');
+const update = require('./update')
+
+module.exports = {
+    getListContacts,
+    getById,
+    add,
+    remove,
+    update,
+}
