+++ models/contacts.json
@@ -6,12 +6,6 @@
     "phone": "(992) 914-3792"
   },
   {
-    "id": "2",
-    "name": "aaaa",
-    "email": "aaa.elementum@scelerisquescelerisquedui.net",
-    "phone": "(000) 206-2688"
-  },
-  {
     "id": "3",
     "name": "Kennedy Lane",
     "email": "mattis.Cras@nonenimMauris.net",
@@ -57,6 +51,12 @@
     "id": "10",
     "name": "Alec Howard",
     "email": "Donec.elementum@scelerisquescelerisquedui.net",
+    "phone": "(748) 206-2688"
+  },
+  {
+    "id": "6341309b335846320b8f5356",
+    "name": "NoFriend",
+    "email": "NoFriend.elementum@scelerisquescelerisquedui.net",
     "phone": "(748) 206-2688"
   }
 ]
