+++ schemas/index.js
@@ -1,0 +1,5 @@
+const addContactSchema = require('./contacts');
+
+module.exports = {
+    addContactSchema,
+}
